requirements.txt
Flask==2.3.0
torch==2.0.0
torchvision==0.15.0
opencv-python==4.7.0.72
requests==2.31.0
matplotlib==3.8.0
Pillow==10.0.0
git+https://github.com/facebookresearch/sam2.git
